Move more code to impl
