+++ ironmq/src/main/scala/akka/stream/alpakka/ironmq/impl/IronMqPullStage.scala
@@ -12,7 +12,7 @@ import akka.stream.alpakka.ironmq._
 import akka.stream.alpakka.ironmq.scaladsl.CommittableMessage
 import akka.stream.stage._
 
-import scala.concurrent.ExecutionContextExecutor
+import scala.concurrent.{ExecutionContextExecutor, Future}
 import scala.util.{Failure, Success}
 
 @InternalApi
@@ -27,9 +27,9 @@ private[ironmq] object IronMqPullStage {
  * This stage will fetch messages from IronMq and buffer them internally.
  *
  * It is implemented as a timed loop, each invocation will fetch new messages from IronMq if the amount of buffered
- * messages is lower than [[ConsumerSettings.bufferMinSize]].
+ * messages is lower than [[akka.stream.alpakka.ironmq.IronMqSettings.ConsumerSettings.bufferMinSize]].
  *
- * The frequency of the loop is controlled by [[ConsumerSettings.fetchInterval]] while the amount of time the client is
+ * The frequency of the loop is controlled by [[akka.stream.alpakka.ironmq.IronMqSettings.ConsumerSettings.fetchInterval]] while the amount of time the client is
  * blocked on the HTTP request waiting for messages is controlled by [[ConsumerSettings.pollTimeout]].
  *
  * Keep in mind that the IronMq time unit is the second, so any value below the second is considered 0.
@@ -49,7 +49,7 @@ private[ironmq] final class IronMqPullStage(queue: Queue.Name, settings: IronMqS
 
   override val shape: SourceShape[CommittableMessage] = SourceShape(out)
 
-  override def createLogic(inheritedAttributes: Attributes): TimerGraphStageLogic =
+  override def createLogic(inheritedAttributes: Attributes): GraphStageLogic =
     new TimerGraphStageLogic(shape) {
 
       implicit def ec: ExecutionContextExecutor = materializer.executionContext
@@ -60,41 +60,30 @@ private[ironmq] final class IronMqPullStage(queue: Queue.Name, settings: IronMqS
       private var buffer: List[ReservedMessage] = List.empty
       private var client: IronMqClient = _ // set in preStart
 
-      override def preStart(): Unit = {
-        super.preStart()
+      override def preStart(): Unit =
         client = IronMqClient(settings)(ActorMaterializerHelper.downcast(materializer).system, materializer)
-      }
-
-      override def postStop(): Unit =
-        super.postStop()
 
       setHandler(
         out,
         new OutHandler {
 
           override def onPull(): Unit = {
-
             if (!isTimerActive(FetchMessagesTimerKey)) {
               schedulePeriodically(FetchMessagesTimerKey, fetchInterval)
             }
-
             deliveryMessages()
           }
         }
       )
 
       override protected def onTimer(timerKey: Any): Unit = timerKey match {
-
         case FetchMessagesTimerKey =>
           fetchMessages()
-
       }
 
       def fetchMessages(): Unit =
         if (!fetching && buffer.size < bufferMinSize) {
-
           fetching = true
-
           client
             .reserveMessages(
               queue,
@@ -116,10 +105,9 @@ private[ironmq] final class IronMqPullStage(queue: Queue.Name, settings: IronMqS
         while (buffer.nonEmpty && isAvailable(out)) {
           val messageToDelivery: ReservedMessage = buffer.head
 
-          val committableMessage = new CommittableMessage {
-            override val message =
-              messageToDelivery.message
-            override def commit() =
+          val committableMessage: CommittableMessage = new CommittableMessage {
+            override val message: Message = messageToDelivery.message
+            override def commit(): Future[Done] =
               client.deleteMessages(queue, messageToDelivery.reservation).map(_ => Done)
           }
