+++ ironmq/src/test/java/akka/stream/alpakka/ironmq/UnitTest.java
@@ -7,6 +7,7 @@ package akka.stream.alpakka.ironmq;
 import akka.actor.ActorSystem;
 import akka.stream.ActorMaterializer;
 import akka.stream.Materializer;
+import akka.stream.alpakka.ironmq.impl.IronMqClient;
 import akka.testkit.javadsl.TestKit;
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
